requirements.txt
aiohttp==3.9.3
asgiref==3.7.2
Django>=4.2,<4.3
sqlparse==0.4.4
tzdata==2023.4
djangorestframework==3.14.0
django-environ==0.11.2
django-extensions==3.2.3
django-modeltranslation==0.18.11
django-countries==7.5.1
psycopg2-binary==2.9.9
pymemcache==4.0.0
python-memcached==1.62
django-debug-toolbar==4.2.0
pytest-django==4.7.0
mixer==7.2.2
pytest-cov==4.1.0
isort==5.13.2
drf-spectacular==0.27.1
requests==2.31.0

black==24.1.1
click==8.1.7
